requirements.txt
altair==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.7.1
beautifultable==0.7.0
bleach==6.0.0
blinker==1.6.2
blis==0.7.11
boto3==1.26.62
botocore==1.29.62
braceexpand==0.1.7
brotlipy==0.7.0
cachetools==5.3.1
catalogue==2.0.10
certifi 
cffi 
cfgv==3.4.0
chardet==4.0.0
charset-normalizer 
confection==0.1.3
contexttimer==0.3.3
contourpy==1.1.1
cryptography 
cycler==0.11.0
cymem==2.0.8
decorator==5.1.1
decord==0.6.0
distlib==0.3.7
docker==3.6.0
docker-pycreds==0.4.0
einops==0.6.1
executing==1.2.0
fairscale==0.4.4
filelock==3.12.4
fonttools==4.42.1
fsspec==2023.9.2
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.37
huggingface-hub==0.17.3
identify==2.5.29
idna 
imageio==2.31.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iopath==0.1.10
ipython==8.12.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kaggle==1.5.16
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
lxml==4.6.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random 
mkl-service==2.4.0
murmurhash==1.0.10
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
numpy
omegaconf==2.3.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow 
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.17.0
portalocker==2.8.2
pre-commit==3.4.0
preshed==3.0.9
prompt-toolkit==3.0.39
protobuf==4.24.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser 
pydantic==2.4.1
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pyOpenSSL 
pyparsing==3.1.1
PySocks 
python-dateutil==2.8.2
python-magic==0.4.27
python-slugify==8.0.1
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests 
rich==13.5.3
rpds-py==0.10.3
s3transfer==0.6.0
safetensors==0.3.3
salesforce-lavis==1.0.2
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six 
smart-open==6.4.0
smmap==5.0.1
soupsieve==2.3.2.post1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.2
streamlit==1.27.0
tenacity==8.2.3
termcolor==1.1.0
text-unidecode==1.3
thinc==8.1.12
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==1.12.1
torchvision==0.13.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
transformers==4.26.1
typer==0.9.0
typing_extensions 
tzdata==2023.3
tzlocal==5.0.1
urllib3 
validators==0.12.6
virtualenv==20.24.5
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.6
webdataset==0.2.57
webencodings==0.5.1
websocket-client==1.5.0
zipp==3.17.0
